requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-docx==1.0.1
PyMuPDF==1.23.4
numpy==1.26.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2